NDRS-763: Extract `State` to a file.
